Lab24: fix

+++ lab24/src/watch.cpp
@@ -162,7 +162,7 @@ watch watch::toClass(const string &sWatch)
 	int counter = 0;
 	unsigned long position = 0;
 	watch tmp;
-	for (int i = 0; i < length; ++i) {
+	for (unsigned long i = 0; i < length; ++i) {
 		if(sWatch[i] == ' '){
 			if (counter == 0){
 				position = clone.find(' ');
@@ -214,7 +214,7 @@ watch watch::toClass(const string &sWatch)
 				clone.erase(0, position + 1);
 				counter++;
 			}else if(counter == 4){
-				for (int j = 0; j < clone.length(); ++j) {
+				for (unsigned long j = 0; j < clone.length(); ++j) {
 					if(clone[j] == ' '){
 						position = clone.find(' ');
 						string firm = clone.substr(0, position);
